Show appro summary in TD preview
